# Faster ObjectType->getEnumCases() type substraction