turns a nested loop into 2 subsequent loops.

----

repro of https://github.com/phpstan/phpstan/issues/11263

before this PR ([c6b961d](https://github.com/phpstan/phpstan-src/commit/c6b961da8cdf6cc12bd0affb2c42da5348074906))
```
time php bin/phpstan analyze slow.php --debug
 8.52s user 0.14s system 99% cpu 8.732 total
time php bin/phpstan analyze slow.php --debug
 8.44s user 0.15s system 99% cpu 8.624 total
```

after this PR
```
time php bin/phpstan analyze slow.php --debug
 7.83s user 0.14s system 99% cpu 8.010 total
time php bin/phpstan analyze slow.php --debug
 7.80s user 0.14s system 99% cpu 7.979 total
```