@thg2k 2024-07-02
There is a typo in the local variable name `subtracedEnumCaseNames`